% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha2prob.R
\name{alpha2prob}
\alias{alpha2prob}
\title{Convert valid alpha array to probabilities under log-linear formulation}
\usage{
alpha2prob(A, J, I)
}
\arguments{
\item{A}{alpha array. Must be valid and provide identifiable log-linear
specification}

\item{J}{index List of variable interactions to include}

\item{I}{vector. Number of categories for each variable in sence}
}
\value{
an array of dimension I with probabilities
}
\description{
Convert valid alpha array to probabilities under log-linear formulation
}
