% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-functions.R
\name{arraySetOnes}
\alias{arraySetOnes}
\title{Initalize array with zeros and ones}
\usage{
arraySetOnes(arrayDim, idx)
}
\arguments{
\item{arrayDim}{Dimension of output array}

\item{idx}{list of vectors specifying location of ones}
}
\value{
array of dimension arrayDim with ones at idx
}
\description{
Initalize array with zeros and ones
}
