% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-functions.R
\name{marginNull2vec}
\alias{marginNull2vec}
\title{Convert Null to full dimension vector}
\usage{
marginNull2vec(mar, A = NULL, mydim = NULL)
}
\arguments{
\item{mar}{list of of margins}

\item{A}{full array used only for dimensionality alternatively pass mydim}

\item{mydim}{dimensionality of full array}
}
\value{
mar with NULL replaced by vectors
}
\description{
Convert Null to full dimension vector
}
