% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-functions.R
\name{marginZero2null}
\alias{marginZero2null}
\title{Convert zero etries of mar to NULL}
\usage{
marginZero2null(mar, A = NULL, mydim = NULL)

marginZero2null(mar, A = NULL, mydim = NULL)
}
\arguments{
\item{mar}{list of of margins}

\item{A}{full array used only for dimensionality alternatively pass mydim}

\item{mydim}{dimensionality of full array}
}
\value{
mar with zero replaced by NULL

mar with zero replaced by NULL
}
\description{
Convert zero etries of mar to NULL

Convert vector ent of mar to NULL
}
