% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-functions.R
\name{marginZero2vec}
\alias{marginZero2vec}
\title{Convert zero entries of mar to full vector}
\usage{
marginZero2vec(mar, mydim = NULL, A = NULL)
}
\arguments{
\item{mar}{list of of margins}

\item{mydim}{dimensionality of full array}

\item{A}{full array used only for dimensionality alternatively pass mydim}
}
\value{
mar with zero replaced by full vector
}
\description{
Convert zero entries of mar to full vector
}
