% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps-allocation.R
\name{query_of_mar}
\alias{query_of_mar}
\title{Check query associated with a given margin}
\usage{
query_of_mar(mar)
}
\arguments{
\item{mar}{list of length 6 with first 1-5 elements being margins
associated with hhgq, sex, cenrace, age, hisp in that order}
}
\value{

}
\description{
Check query associated with a given margin
}
