% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps-allocation.R
\name{queryChar2queryMod}
\alias{queryChar2queryMod}
\title{Convert query character to epsilon modifier}
\usage{
queryChar2queryMod(queryMod, query)
}
\arguments{
\item{queryMod}{vector of modifier levels of query}

\item{query}{character valued geography}
}
\value{
entry of geoMod associated with geo
}
\description{
Convert query character to epsilon modifier
}
