% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-functions.R
\name{recode}
\alias{recode}
\title{Recode or condense a marginal of histogram}
\usage{
recode(A, margin, newCode)
}
\arguments{
\item{A}{array to be recoded}

\item{margin}{margin that will be recoded}

\item{newCode}{list of vectors that specify the recode}
}
\value{
recoded array
}
\description{
Recode or condense a marginal of histogram
}
